src/requirements.txt
Coffin==0.3.7
Django==1.4.3
django-appconf==1.0.1
django-compressor==1.4
django-jasmine==0.4.1
Jinja2==2.6
MarkupSafe==1.0
psycopg2-binary==2.8.4
Pygments==2.4.2
six==1.12.0
xlwt==0.7.5
django-user-agents==0.4.0
gunicorn==19.10.0
python-dotenv==0.18.0
